sniff function bug gix
