+++ requirements.txt
@@ -1,4 +1,5 @@
 fastapi
+netifaces
 uvicorn
 scapy
 libpcap
